fix: make openInterest optional in event deserialization
The Polymarket API omits openInterest on some events (e.g. 15m SOL).
Adding #[serde(default)] lets the field fall back to 0.0 instead of
failing with a missing-field error.

+++ src/models/event.rs
@@ -26,7 +26,8 @@ pub struct PolyResponseEvent {
     pub restricted: bool,
     //liquidity: f64,
     //volume: f64,
-    pub open_interest: i64,
+    #[serde(default)]
+    pub open_interest: f64,
     pub created_at: String,
     pub updated_at: String,
     //competitive: f64,
